Namespace tidy-up - CombinedFile
