+++ src/DigitalPreservation/DigitalPreservation.Common.Model/DepositHelpers/MinimalItem.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 using DigitalPreservation.Common.Model.Transit;
+using DigitalPreservation.Common.Model.Transit.Combined;
 
 namespace DigitalPreservation.Common.Model.DepositHelpers;
