+++ src/DigitalPreservation/DigitalPreservation.UI/Pages/Deposits/_RenderCombinedDirectoryAsTableRows.cshtml
@@ -1,5 +1,6 @@
 ﻿@using DigitalPreservation.Common.Model
 @using DigitalPreservation.Common.Model.Transit
+@using DigitalPreservation.Common.Model.Transit.Combined
 @using DigitalPreservation.Common.Model.Transit.Extensions.Metadata
 @using DigitalPreservation.Utils
 @using Storage.Repository.Common
