+++ src/DigitalPreservation/DigitalPreservation.Workspace/Requests/DeleteItems.cs
@@ -7,6 +7,7 @@ using DigitalPreservation.Common.Model.DepositHelpers;
 using DigitalPreservation.Common.Model.Mets;
 using DigitalPreservation.Common.Model.Results;
 using DigitalPreservation.Common.Model.Transit;
+using DigitalPreservation.Common.Model.Transit.Combined;
 using DigitalPreservation.Utils;
 using MediatR;
 using Microsoft.Extensions.Logging;
